make tx validate input type more explicit, MIN/MAX values of trit/trytes in bytes.

+++ src/main/java/com/iota/iri/hash/ISSInPlace.java
@@ -13,8 +13,8 @@ public class ISSInPlace {
   public static final int TRYTE_WIDTH = 3;
   private static final int NUMBER_OF_SECURITY_LEVELS = 3;
   public static final int NORMALIZED_FRAGMENT_LENGTH = Kerl.HASH_LENGTH / TRYTE_WIDTH / NUMBER_OF_SECURITY_LEVELS;
-  private static final int MIN_TRIT_VALUE = -1, MAX_TRIT_VALUE = 1;
-  private static final int MIN_TRYTE_VALUE = -13, MAX_TRYTE_VALUE = 13;
+  private static final byte MIN_TRIT_VALUE = -1, MAX_TRIT_VALUE = 1;
+  private static final byte MIN_TRYTE_VALUE = -13, MAX_TRYTE_VALUE = 13;
 
   public static void subseed(SpongeFactory.Mode mode, byte[] subseed, int index) {
